Convert `themeHandeler` function to a hook Convert `themeHandeler` function to a hook with name `useCustomTheme`
- Move `appStyle` that was exist in `App` component into `useCustomTheme` hook.

- Import `appStyle` and theme object from `useCustomTheme` hook in `App` component
to use them for style the app

- Remove `SwitchThemeContext` due to `useCookies` hook is enough to handle
switching theme process.

- Inside `SwitchTheme` component remove the usage of `SwitchThemeContextin`
and make some changes in the styles of `MaterialUISwitch` component.

- Disable `StrictMode` component in `main.tsx` component temporarily
