+++ src/App.tsx
@@ -4,15 +4,13 @@ import { createContext } from "react";
 import AdminAppBar from "./components/AdminBar";
 import { Box, ThemeProvider } from "@mui/material";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import themeHandeler from "./functions/themeHandeler";
 import { ToastContainer } from 'react-toastify';
-import { useCookies } from 'react-cookie';
 import useAdminLogIn from "./hooks/useAdminLogIn";
 import LogInForm from "./components/LogInForm";
 import { ErrorThrower, LoadingCircle, LoadingPage } from "@abdulrhmangoni/am-store-library";
 import { Outlet } from "react-router-dom";
 import { AdminData } from "./types/dataTypes";
-import SwitchThemeContext from "./components/SwitchThemeContext";
+import useCustomTheme from "./hooks/useCustomTheme";
 
 export const AdminDataContext = createContext<AdminData | null>(null);
 
@@ -26,47 +24,29 @@ export default function App() {
     isNetworkError, isLogged,
     isOut, isServerError
   } = useAdminLogIn();
-  const theme = themeHandeler(useCookies()[0].theme ?? "dark");
-  const { palette: { primary, background, text } } = theme
 
-  const AppStyle = {
-    display: "flex",
-    flexDirection: "column",
-    minHeight: "100vh",
-    "& *::-webkit-scrollbar-thumb": { bgcolor: "primary.main" },
-    "& *::-webkit-scrollbar": { bgcolor: "white", width: "3px", height: "6px" },
-    "& input:autofill": {
-      boxShadow: `0 0 0 100px ${background.default} inset !important`,
-      WebkitTextFillColor: `${text.primary} !important`
-    },
-    ":root": {
-      "--toastify-color-info": primary.main,
-      "--toastify-color-dark": background.default
-    }
-  }
+  const { theme, appStyle } = useCustomTheme();
 
   return (
-    <AdminDataContext.Provider value={adminData}>
-      <SwitchThemeContext>
-        <ThemeProvider theme={theme}>
-          <QueryClientProvider client={queryClient}>
-            <Box sx={AppStyle}>
-              {
-                isLoading ? <LoadingPage />
-                  : isLogged ? <AppContent />
-                    : isOut ? <LogInForm />
-                      : isNetworkError ? <NetworkError />
-                        : isServerError ? <ServerError />
-                          : isError ? <UnexpectedError />
-                            : null
-              }
-            </Box>
-            <LoadingCircle staticCircle darkBg />
-            <ToastContainer limit={4} position="bottom-left" theme="colored" />
-          </QueryClientProvider>
-        </ThemeProvider>
-      </SwitchThemeContext>
-    </AdminDataContext.Provider>
+    <ThemeProvider theme={theme}>
+      <AdminDataContext.Provider value={adminData}>
+        <QueryClientProvider client={queryClient}>
+          <Box sx={appStyle}>
+            {
+              isLoading ? <LoadingPage />
+                : isLogged ? <AppContent />
+                  : isOut ? <LogInForm />
+                    : isNetworkError ? <NetworkError />
+                      : isServerError ? <ServerError />
+                        : isError ? <UnexpectedError />
+                          : null
+            }
+          </Box>
+          <LoadingCircle staticCircle darkBg />
+          <ToastContainer limit={4} position="bottom-left" theme="colored" />
+        </QueryClientProvider>
+      </AdminDataContext.Provider>
+    </ThemeProvider>
   )
 }
